Incertus.jl

The package implements several restricted randomization procedures for two- and multi-arm clinical trials, targeting equal or unequal allocation.

Randomization, targeting equal 1:1 allocation

The following settings are assumed:

There are two treatment arms investigated in a trial:

$E$ is an experimental treatment arm.

$C$ is a control treatment arm.

$n$ is a total sample size.

$N_1(j)$ and $N_2(j)$ are treatment numbers, i.e., sample sizes on treatments after the $j^\text{th}$ allocation step ($N_1(j)+N_2(j) = j, \: j = 1, \ldots, n$).

$\delta_j$ is a treatment indicator:

\[\delta_j = \left\{\begin{array}{rl} 1, &\text{if treatment }E \\ 0, &\text{if treatment }C \end{array}\right. .\]

Under these assumptions, a restricted randomization procedure is defined as

\[\begin{aligned} \phi_1 &= \Pr(\delta_1 = 1) = 0.5; \\ \phi_j &= \Pr(\delta_j = 1|\delta_1, \ldots, \delta_{j-1}), \: j = 2, \ldots, n. \end{aligned}\]

Randomization, targeting unequal allocation in two-arm trial and equal/unequal allocation in multi-arm trial

The following settings are assumed:

$K$ is a number of treatment arms in a trial ($K \geq 2$).

$\mathbf{w} = \left(w_1,\ldots, w_K\right)'$ is the fixed allocation ratio, $w_1:\ldots :w_K$, where $w_k$'s are positive, not necessarily equal numbers (usually, integers) with the greatest common divisor of 1.

$\boldsymbol{\rho} = \left(\rho_1, \ldots, \rho_K\right)'$ is a vector of target allocation proportions, where

\[\rho_k = \frac{w_k}{\sum_{k=1}^{K}{w_k}}, 0 \leq \rho_k \leq 1, \text{ and }\sum_{k=1}^K{\rho_k} = 1.\]

$n$ is a total sample size.

$\mathbf{N}(j) = \left(N_1(j), \ldots, N_K(j)\right)'$ is a vector of treatment numbers, i.e., numbers of subjects allocated to $K$ treatments after $j$ allocations ($1 \leq j \leq n$).

Note that, in general, $N_k(j)$'s are random variables with $\sum_{k = 1}^K{N_k(j)} = j$.

$\mathbf{P}(j) = \left(P_1(j), \ldots, P_K(j)\right)'$ is a vector of randomization (allocation) probabilities for the $j^\text{th}$subject.

Note that $0 \leq P_k(j) \leq 1$, and $\sum_{k = 1}^K{P_k(j)} = 1$ for each $j = 1, 2, \ldots, n$.

Also, note that in general, $\mathbf{P}(j)$ depends on $\mathbf{N}(j-1)$ (in generalization of Efron's BCD) or on $\frac{\mathbf{N}(j-1)}{j-1}$ (in generalization of Wei's urn design).

Under these assumptions, a restricted randomization procedure is defined as

\[\begin{aligned} P_k(1) &= \Pr(\delta_1 = k) = \rho_k, \: k = 1, \ldots, K; \\ P_k(j) &= \Pr(\delta_j = k|\delta_1, \ldots, \delta_{j-1}), \: k = 1, \ldots, K; \: j = 2, \ldots, n. \end{aligned}\]

Procedures implemented in the package

Completely Randomized Design (CRD)

For a two-arm trial and 1:1 target allocation, every subject is allocated to treatments with a fixed probability

\[\phi_j = 0.5, \: j = 1, \ldots, n.\]

In case of two-arm trial with unequal allocation or multi-arm trial with equal/unequal allocation, every subject is allocated to treatments with fixed probabilities that are equal to the target allocation proportions:

\[P_k(j) = \rho_k, \: k = 1, \ldots, K; \: j = 1, \ldots, n.\]

Incertus.CRD — Type

A type of randomization, representing Completely Randomized Design (CRD).

CRD() command initializes a complete randomization procedure, targeting 1:1 allocation.

CRD(w) command initializes a complete randomization procedure, targeting allocation specified by w.

An output of both commands is an instance of CRD.

julia> using Incertus
julia> crd =  CRD()     # complete randomization, targeting 1:1 allocationCRD: complete randomization procedure, targeting 1:1 allocation in 2-arm trial.
julia> w = [1, 2, 3, 4]4-element Vector{Int64}:
 1
 2
 3
 4
julia> crd =  CRD(w)    # complete randomization, targeting allocation specified by wCRD: complete randomization procedure, targeting 1:2:3:4 allocation in 4-arm trial.

Permuted Block Design (PBD)

Treatment assignments are made in blocks of size $bs$ (for a two-arm trial and 1:1 target allocation, $bs= 2\lambda$; otherwise, $bs=\lambda W$, where $W=w_1 + \ldots + w_K$ is a sum of elements of vector $\mathbf{w}$, target allocation vector. Here, $\lambda$ is a parameter of the PBD, representing the number of minimal balanced sets in the block of size $bs$).

At the $j^\text{th}$ allocation step, let $k^{(j-1)} = \left \lfloor \frac{j-1}{bs}\right\rfloor$ ($\lfloor x \rfloor$ is a floor function that returns the greatest integer less than or equal to $x$). In essence, $k^{(j-1)}$ is the number of complete blocks among the first $j-1$ assignments.

The probabilities of treatment assignments within each block are changed according the current imbalance in a block:

for a two-arm trial and 1:1 target allocation,

\[\phi_j = \frac{0.5bs(1+k^{(j-1)})-N_1(j-1)}{bs(1+k^{(j-1)})-(j-1)}, \: j = 1, \ldots, n;\]

for a two-arm trial with unequal allocation or multi-arm trial with equal/unequal allocation,

\[P_k(j) = \frac{w_k\lambda(1+k^{(j-1)})-N_k(j-1)}{bs(1+k^{(j-1)})-(j-1)}, k = 1, \ldots, K; \: j = 1, \ldots, n.\]

See [Zhao and Weng (2011), page 955, equation (5)]

Incertus.PBD — Type

A type of rectricted randomization, representing Permuted Block Design (PBD).

PBD(λ) command initializes a permuted block design with a block size equal to 2λ, targeting 1:1 allocation.

PBD(w, λ) command initializes a permuted block design with a parameter λ, targeting allocation specified by w; block size equals to λ*sum(w).

An output of both commands is an instance of PBD.

julia> using Incertus
julia> pbd = PBD(1)      # PBD, targeting 1:1 allocation, with a block size equal to 2*1 = 2PBD(1): restricted randomization procedure, targeting 1:1 allocation in 2-arm trial.
julia> pbd = PBD(3)      # PBD, targeting 1:1 allocation, with a block size equal to 2*3 = 6PBD(3): restricted randomization procedure, targeting 1:1 allocation in 2-arm trial.
julia> w = [1, 2, 3, 4]4-element Vector{Int64}:
 1
 2
 3
 4
julia> pbd =  PBD(w, 1)  # PBD, targeting allocation specified by w, with a block size sum(w)PBD(1): restricted randomization procedure, targeting 1:2:3:4 allocation in 4-arm trial.
julia> pbd =  PBD(w, 3)  # PBD, targeting allocation specified by w, with a block size 3*sum(w)PBD(3): restricted randomization procedure, targeting 1:2:3:4 allocation in 4-arm trial.

Random Allocation Rule (Rand)

A version of PBD, when the block size $bs$ equals to the total sample size $n$. At the $j^\text{th}$ allocation step, probabilities of treatment assignments are calculated as:

for a two-arm trial and 1:1 target allocation,

\[\phi_j = \frac{0.5n-N_1(j-1)}{n-(j-1)}, \: j = 1, \ldots, n;\]

for a two-arm trial with unequal allocation or multi-arm trial with equal/unequal allocation,

\[P_k(j) = \frac{nw_k/W-N_k(j-1)}{n-(j-1)}, k = 1, \ldots, K; \: j = 1, \ldots, n,\]

where $W=w_1 + \ldots + w_K$ is a sum of elements of vector $\mathbf{w}$, target allocation vector.

Incertus.RAND — Type

A type of rectricted randomization, representing Random Allocation Rule (Rand).

RAND(n) command initializes a random allocation rule , targeting 1:1 allocation in a trial with a sample size equal to n.

RAND(w, n) command initializes a random allocation rule , targeting allocation specified by w in a trial with a sample size equal to n.

An output of both commands is an instance of RAND.

julia> using Incertus
julia> rnd = RAND(50)     # RAND, targeting 1:1 allocation, in a trial with 50 subjectsRand: restricted randomization procedure, targeting 1:1 allocation in 2-arm trial.
julia> w = [1, 2, 3, 4]4-element Vector{Int64}:
 1
 2
 3
 4
julia> rnd =  RAND(w, 50) # RAND, targeting allocation specified by w, in a trial with 50 subjectsRand: restricted randomization procedure, targeting 1:2:3:4 allocation in 4-arm trial.

Truncated Binomial Design (TBD)

Treatment assignments are made with probability 0.5 until one of the treatments receives its quota of $\frac{n}{2}$ subjects; thereafter all remaining assignments are made deterministically to the opposite treatment.

At the $j^\text{th}$allocation step, let $N_1$ and $N_2$ be the numbers of subjects allocated to treatments s.t. $N_1+N_2 = j-1.$ Then,

\[\phi_j = \left\{\begin{array}{rl} 0.5, & \max(N_1, N_2) < \frac{n}{2} \\ 1, & N_1 < N2 \\ 0, & N_1 > N2 \end{array}\right. ,\: j = 1, \ldots, n.\]

Incertus.TBD — Type

A type of rectricted randomization, representing Truncated Binomial Design (TBD).

TBD(n) command initializes a truncated binomial design, targeting 1:1 allocation in a trial with a sample size equal to n.

An output of the command is an isntance of TBD.

julia> using Incertus
julia> tbd = TBD(50)  # TBD, targeting 1:1 allocation, in a trial with 50 subjectsTBD: restricted randomization procedure, targeting 1:1 allocation in 2-arm trial.

Efron's Biased Coin Design (EBCD)

At any allocation step, if treatment numbers $N_1$ and $N_2$ are balanced, the next assignment is made with probability 0.5; otherwise, the underrepresented treatment is assigned with probability $p$, where $0.5 \leq p \leq 1$ is a fixed and pre-specified parameter that determines the trade-off between balance and randomness.

At the $j^\text{th}$ allocation step, given treatment numbers $N_1$ and $N_2$, s.t. $N_1+N_2 = j-1$, and imbalance $d = N_1-N_2$,

\[\phi_j = \left\{\begin{array}{rl} 0.5, & N_1 = N_2\\ p, & N_1 < N2 \\ 1-p, & N_1 > N2 \end{array}\right. ,\: j = 1, \ldots, n.\]

Note that $p=1$ corresponds to PBD with block size $b=2$.

Incertus.EBCD — Type

A type of rectricted randomization, representing Efron's Biased Coin Design (EBCD).

EBCD(p) command initializes Efron's biased coin design with a parameter equal to p, targeting 1:1 allocation.

An output of the command is an isntance of EBCD.

julia> using Incertus
julia> ebcd = EBCD(2//3) # EBCD, targeting 1:1 allocation, with parameter p=2/3EBCD(2/3): restricted randomization procedure, targeting 1:1 allocation in 2-arm trial.

Adjustable Biased Coin Design (ABCD)

An extension of Efron’s BCD. At the $j^\text{th}$ allocation step, given treatment numbers $N_1$ and $N_2$, s.t. $N_1+N_2 = j-1$, and imbalance $d = N_1-N_2$,

\[\phi_j = \left\{\begin{array}{rl} 0.5, & |d| <= 1 \\ \frac{|d|^a}{1+|d|^a}, & d < -1 \\ \frac{1}{1+|d|^a}, & d > 1 \end{array}\right. ,\: j = 1, \ldots, n.\]

Incertus.ABCD — Type

A type of rectricted randomization, representing Adjustable Biased Coin Design (ABCD).

ABCD(a) command initializes adjustable biased coin design with a parameter equal to a, targeting 1:1 allocation.

An output of the command is an isntance of ABCD.

julia> using Incertus
julia> abcd = ABCD(2) # ABCD, targeting 1:1 allocation, with parameter a=2ABCD(2): restricted randomization procedure, targeting 1:1 allocation in 2-arm trial.

Generalized Biased Coin Design (GBCD)

A generalization of Efron’s BCD. At the $j^\text{th}$ allocation step, given treatment numbers $N_1$ and $N_2$, s.t. $N_1+N_2 = j-1$, and imbalance $d = N_1-N_2$,

\[\phi_j = \left\{\begin{array}{rl} 0.5, & j = 1 \\ \frac{N_2^\gamma}{N_1^\gamma+N_2^\gamma}, & j = 1, \ldots, n. \end{array}\right. \]

Incertus.GBCD — Type

A type of rectricted randomization, representing Generalized Biased Coin Design (GBCD).

GBCD(γ) command initializes generalized biased coin design with a parameter equal to γ, targeting 1:1 allocation.

An output of the command is an isntance of GBCD.

julia> using Incertus
julia> gbcd = GBCD(2) # GBCD, targeting 1:1 allocation, with parameter γ=2GBCD(2): restricted randomization procedure, targeting 1:1 allocation in 2-arm trial.

Big Stick Design (BSD)

An example of maximum tolerated imbalance (MTI) procedures. It makes prediction of the future treatment allocations more difficult (even knowing the current sizes of the treatment groups) and controls treatment imbalance at a predefined threshold throughout the experiment. A general MTI procedure specifies a certain boundary for treatment imbalance, say $mti$, that cannot be exceeded.

At the $j^\text{th}$ allocation step, given treatment numbers $N_1$ and $N_2$, s.t. $N_1+N_2 = j-1$, and imbalance $d = N_1-N_2$,

\[\phi_j = \left\{\begin{array}{rl} 0.5, & |d| < mti \\ 0, & d = mti \\ 1, & d = -mti \end{array}\right. ,\: j = 1, \ldots, n.\]

Incertus.BSD — Type

A type of rectricted randomization, representing Big Stick Design (BSD).

BSD(mti) command initializes big stick design with a parameter equal to mti, targeting 1:1 allocation.

An output of the command is an isntance of BSD.

julia> using Incertus
julia> bsd = BSD(3) # BSD, targeting 1:1 allocation, with parameter mti=3BSD(3): restricted randomization procedure, targeting 1:1 allocation in 2-arm trial.

Biased Coin Design With Imbalance Tolerance (BCDWIT)

A combination of Efron’s BCD and BSD. At the $j^\text{th}$ allocation step, given treatment numbers $N_1$ and $N_2$, s.t. $N_1+N_2 = j-1$, and imbalance $d = N_1-N_2$,

\[\phi_j = \left\{\begin{array}{rl} 0.5, & |d| < mti\: \& \: d = 0 \\ p, & |d| < mti \: \& \: d < 0 \\ 1-p, & |d| < mti \: \& \: d > 0 \\ 0, & d = mti \\ 1, & d = -mti \end{array}\right. ,\: j = 1, \ldots, n.\]

Incertus.BCDWIT — Type

A type of rectricted randomization, representing Biased Coin Design With Imbalance Tolerance (BCDWIT).

BCDWIT(p, mti) command initializes biased coin design with imbalance tolerance with parameters equal to p and mti, targeting 1:1 allocation.

An output of the command is an isntance of BCDWIT.

julia> using Incertus
julia> bcdwit = BCDWIT(2//3, 3) # BCDWIT with p = 2/3 and mti=3, targeting 1:1 allocationBCDWIT(2/3, 3): restricted randomization procedure, targeting 1:1 allocation in 2-arm trial.

Block Urn Design (BUD)

This design was proposed by Zhao and Weng (2011), to provide a more random design than the PBD. Let $N_k(j-1)$ denote the number of treatment $k$ assignments among first $j-1$ subjects, and $k^{(j-1)}=\min\limits_{1 \leq k \leq K} \left \lfloor \frac{N_k(j-1)}{w_k}\right \rfloor$ denote the number of minimal balanced sets among the first $j-1$ assignments. Then, at the $j^\text{th}$ allocation step, probabilities of treatment assignments are calculated as:

for a two-arm trial and 1:1 target allocation,

\[\phi_j = \frac{\lambda+\min(N_1(j-1), N_2(j-1))-N_1(j-1)}{2(\lambda+\min(N_1(j-1), N_2(j-1)))-(j-1)}, \: j = 1, \ldots, n;\]

for a two-arm trial with unequal allocation or multi-arm trial with equal/unequal allocation,

\[P_k(j) = \frac{w_k(\lambda+k^{(j-1)})-N_k(j-1)}{W(\lambda+k^{(j-1)})-(j-1)}, k = 1, \ldots, K; \: j = 1, \ldots, n,\]

where $W=w_1 + \ldots + w_K$ is a sum of elements of vector $\mathbf{w}$, target allocation vector.

See [Zhao and Weng (2011), page 955, equations (2) and (3)].

Incertus.BUD — Type

A type of rectricted randomzation, representing Block Urn Design (BUD).

BUD(λ) command initializes block urn desin with a parameter equal to $λ$, targeting 1:1 allocation.

BUD(w, λ) command initializes block urn desin with a parameter equal to $λ$, targeting allocation specified by w.

An output of both commands is an instance of BUD.

julia> using Incertus
julia> bud = BUD(2)      # BUD, targeting 1:1 allocation (λ=2)BUD(2): restricted randomization procedure, targeting 1:1 allocation in 2-arm trial.
julia> w = [1, 2, 3, 4]4-element Vector{Int64}:
 1
 2
 3
 4
julia> bud =  BUD(w, 2)  # BUD, targeting allocation specified by w (λ=2)BUD(2): restricted randomization procedure, targeting 1:2:3:4 allocation in 4-arm trial.

Ehrenfest Urn Design (EUD)

Another example of the maximum tolerated imbalance (MTI) procedure. At the $j^\text{th}$ allocation step, given treatment numbers $N_1$ and $N_2$, s.t. $N_1+N_2 = j-1$, and imbalance $d = N_1-N_2$,

\[\phi_j = \frac{1}{2}\left(1-\frac{d}{mti}\right) ,\: j = 1, \ldots, n,\]

where $mti$ ($>0$) is a parameter of the procedure.

Incertus.EUD — Type

A type of rectricted randomzation, representing Ehrenfest Urn Design (BUD).

EUD(mti) command initializes Ehrenfest's urn design with a parameter equal to $mti$, targeting 1:1 allocation.

An output of the command is an isntance of EUD.

julia> using Incertus
julia> eud = EUD(2) # EUD, targeting 1:1 allocation (mti=2)EUD(2): restricted randomization procedure, targeting 1:1 allocation in 2-arm trial.

Bayesian Biased Coin Design (BBCD)

A special class of biased coind designs (BCDs). At the $j^\text{th}$ allocation step, given treatment numbers $N_1$ and $N_2$, s.t. $N_1+N_2 = j-1$,

\[\phi_j = \left\{\begin{array}{rl} 0.5, & j = 1 \\ 1, & j = 2 \: \& \: N_1 = 0 \\ 0, & j = 2 \: \& \: N_1 = 1 \\ \frac{\left(1 + \frac{N_2}{nN_1}\right)^\frac{1}{\gamma}}{\left(1 + \frac{N_2}{nN_1}\right)^\frac{1}{\gamma} + \left(1 + \frac{N_1}{nN_2}\right)^\frac{1}{\gamma}}, & j \geq 3 \end{array}\right. ,\: j = 1, \ldots, n.\]

where $\gamma$ ($>0$) is a parameter of the procedure, and $n$ is a sample size (pre-specified).

Incertus.BBCD — Type

A type of rectricted randomzation, representing Bayesian Biased Coin Design (BBCD).

BBCD(γ, n) command initializes Bayesian biased coin design with a parameter equal to $γ$, targeting 1:1 allocation in a trial with a sample size equal to n.

An output of the command is an isntance of BBCD.

julia> using Incertus
julia> bbcd = BBCD(0.1, 40) # BBCD, targeting 1:1 allocation (γ=2, n = 40)BBCD(0.1): restricted randomization procedure, targeting 1:1 allocation in 2-arm trial.

Funcions implemented to calculate allocation probabilities

Incertus.allocation_prb — Method

Function calculates allocation probabilities for CRD.

Call

allocation_prb(rnd)

Arguments

rnd::CRD: an object, representing Complete Randomization.

Incertus.allocation_prb — Method

Function calculates allocation probabilities for PBD, given treatment numbers.

Call

allocation_prb(rnd, N)

Arguments

rnd::PBD: an object, representing Permuted Block Design.

N::Vector{Int64}: a vector of current treatment numbers.

Incertus.allocation_prb — Method

Function calculates allocation probabilities for RAND, given treatment numbers.

Call

allocation_prb(rnd, N)

Arguments

rnd::RAND: an object, representing Random Allocation Rule.

N::Vector{Int64}: a vector of current treatment numbers.

Incertus.allocation_prb — Method

Function calculates allocation probabilities for TBD, given treatment numbers.

Call

allocation_prb(rnd, N)

Arguments

rnd::TBD: an object, representing Truncated Binomial Design.

N::Vector{Int64}: a vector of current treatment numbers.

Incertus.allocation_prb — Method

Function calculates allocation probabilities for EBCD, given treatment numbers.

Call

allocation_prb(rnd, N)

Arguments

rnd::EBCD: an object, representing Efron's Biased Coin Design.

N::Vector{Int64}: a vector of current treatment numbers.

Incertus.allocation_prb — Method

Function calculates allocation probabilities for ABCD, given treatment numbers.

Call

allocation_prb(rnd, N)

Arguments

rnd::ABCD: an object, representing Adjustable Biased Coin Design.

N::Vector{Int64}: a vector of current treatment numbers.

Incertus.allocation_prb — Method

Function calculates allocation probabilities for GBCD, given treatment numbers.

Call

allocation_prb(rnd, N)

Arguments

rnd::GBCD: an object, representing Generalized Biased Coin Design.

N::Vector{Int64}: a vector of current treatment numbers.

Incertus.allocation_prb — Method

Function calculates allocation probabilities for BSD, given treatment numbers.

Call

allocation_prb(rnd, N)

Arguments

rnd::BSD: an object, representing Big Stick Design.

N::Vector{Int64}: a vector of current treatment numbers.

Incertus.allocation_prb — Method

Function calculates allocation probabilities for BCDWIT, given treatment numbers.

Call

allocation_prb(rnd, N)

Arguments

rnd::BCDWIT: an object, representing Biased Coin Design With Imbalance Tolerance.

N::Vector{Int64}: a vector of current treatment numbers.

Incertus.allocation_prb — Method

Function calculates allocation probabilities for BUD, given treatment numbers.

Call

allocation_prb(rnd, N)

Arguments

rnd::BUD: an object, representing Block Urn Design.

N::Vector{Int64}: a vector of current treatment numbers.

Incertus.allocation_prb — Method

Function calculates allocation probabilities for EUD, given treatment numbers.

Call

allocation_prb(rnd, N)

Arguments

rnd::EUD: an object, representing Ehrenfest Urn Design.

N::Vector{Int64}: a vector of current treatment numbers.

Incertus.allocation_prb — Method

Function calculates allocation probabilities for BBCD, given treatment numbers.

Call

allocation_prb(rnd, N)

Arguments

rnd::BBCD: an object, representing Bayesian Biased Coin Design.

N::Vector{Int64}: a vector of current treatment numbers.

Simulation

To perform simulations, the following functionality has been implemented:

Incertus.SimulatedRandomization — Type

A type, representing an output of a simulated randomization procedure.

A command SimulatedRandomization(label, trt, prb) initializes an instance of SimulatedRandomization:

label is a string that describes the simulated randomization procedure;

trt is a matrix of size $n\times S$, representing treatment assignments;

prb is an array of size $n\times K\times S$, representing allocation probabilities,

where

$n$ is the number of subjects simulated;

$K$ is the number of treatments simulated (equal to the length of the target allocation vector);

$S$ is the number of simulations performed.

Incertus.simulate — Function

Function simulates a randomization procedure

Call:

simulate(rnd, nsbj, nsim, seed)

Arguments

rnd::Randomization: an object, representing a randomization procedure to be simulated.

nsbj::Int64: number of subjects simulated.

nsim::Int64: number of simulations performed.

seed::Int64: a random seed (for reproducibility); a default is set to 314159.

Result

an object, representing simlated randomization, an instance of SimulatedRandomization.

Incertus.simulate — Function

Function simulates a set of randomization procedures

Call:

simulate(rnd, nsbj, nsim, seed)

Arguments

rnd::Vector{Randomization}: a vector of instances of Randomization type; each instance represents a randomization procedure to be simulated.

nsbj::Int64: number of subjects simulated.

nsim::Int64: number of simulations performed.

seed::Int64: a random seed (for reproducibility); a default is set to 314159.

Result

a vector of instances of SimulatedRandomization type; each instance represents a simulated randomization procedure.

Operational characteristics (two-arm trial with equal 1:1 allocation)

Several measures of imbalance and randomness have been implemented in the package:

Measures of imbalance:

$\mathbf{E}\left[|D(j)|\right]$ – expected absolute imbalance at the $j^\text{th}$ allocation step.

$\mathbf{E}\left[|D(j)|^2\right]=\mathbf{var}\left[D(j)\right]$ – variance of imbalance at the $j^\text{th}$ allocation step.

$\mathbf{E}\left[\max\limits_{1\leq m \leq j}|D(m)|\right]$ – expected maximum imbalance over the first $j$ allocation steps.

$Imb(j) = \frac{1}{j}\sum\limits_{m=1}^j\frac{\mathbf{E}\left[|D(m)|^2\right]}{m}$ – a cumulative average loss at the $j^\text{th}$ allocation step.

Measures of randomness:

$EPCG_{conv}(j) = \frac{1}{j}\sum\limits_{m=1}^j\mathbf{E}\left[G_m = \delta_m\right]$ – cumulative average of expected proportions of correct guesses over first ($j$) allocation steps under the convergence guessing strategy, where $G_m$ is a random variable taking values based on the investigator's guess (if $G_m = \delta_m$, the guess is correct) at the $m^\text{th}$ allocation step, given current imbalance, $D(m-1)$:

\[G_m = \left\{ \begin{array}{rl} 1, & D(m-1) < 0; \\ \sim Bernoulli(0.5), & D(m-1) = 0; \\ 0, & D(m-1) > 0. \end{array} \right. \]

$EPCG_{max}(j) = \frac{1}{j}\sum\limits_{m=1}^j\mathbf{E}\left[\widetilde{G}_m = \delta_m\right]$ – cumulative average of expected proportions of correct guesses over first ($j$) allocation steps under the maximum probability guessing strategy, where $\widetilde{G}_m$ is a random variable taking values based on the investigator's guess (if $G_m = \delta_m$, the guess is correct) at the $m^\text{th}$ allocation step, given allocation probability, $\phi_{m}$:

\[\widetilde{G}_m = \left\{ \begin{array}{rl} 1, & \phi_m > 0.5; \\ \sim Bernoulli(0.5), & \phi_m = 0.5; \\ 0, & \phi_m < 0.5. \end{array} \right. \]

$PD(j) = \frac{1}{j}\sum\limits_{m=1}^j\Pr(\phi_m\in\{0, 1\})$ – cumulative average of expected proportions of deterministic assignments over first ($j$) allocation steps.

$FI(j) = \frac{4}{j}\sum\limits_{m=1}^j\mathbf{E}\left[|\phi_m-0.5|\right]$ – forcing index, which takes values on a scale 0–1. $FI(j) = 0, \forall j$ for CRD and $FI(j) = 1$ for PBD with a block size $bs=2$, assuming $j$ is even (most balanced design).

Balance-randomness trade-off:

$G(j) = \sqrt{\left\{Imb(j)\right\}^2 + \left\{FI(j)\right\}^2}$ represents a balance-randomness trade-off at the $j^\text{th}$ allocation step. Lower values of $G(j)$ indicate better balance–randomness trade-off.

Below, there are functions available for calculating operational characteristics.

Final imbalance

Incertus.calc_final_imb — Method

Function calculates final imbalance

Call

calc_final_imb(sr)

Arguments

sr::SimulatedRandomization: an instance of SimulatedRandomization, an object, representing simulation output.

Result

A vector of final imbalance values obtained via simulations.

Incertus.calc_final_imb — Method

Function calculates final imbalance

Call

calc_final_imb(sr)

Arguments

sr::Vector{SimulatedRandomization}: a vector of instances of SimulatedRandomization, representing simulation output.

Result

A DataFrame of simulated final imbalances' values obtained via simulations.

Expected absolute imbalance

Incertus.calc_expected_abs_imb — Method

Function calculates expected absolute imbalance vs. allocation step.

Call

calc_expected_abs_imb(sr)

Arguments

sr::SimulatedRandomization: an instance of SimulatedRandomization, an object, representing simulation output.

Result

A vector of expected absolute imbalance values summarized via simulations.

Incertus.calc_expected_abs_imb — Method

Function calculates expected absolute imbalance vs. allocation step.

Call

calc_expected_abs_imb(sr)

Arguments

sr::Vector{SimulatedRandomization}: a vector of instances of SimulatedRandomization, representing simulation output.

Result

A DataFrame of expected absolute imbalances' values summarized via simulations.

Variance of imbalance

Incertus.calc_variance_of_imb — Method

Function calculates variance of imbalance vs. allocation step.

Call

calc_variance_of_imb(sr)

Arguments

sr::SimulatedRandomization: an instance of SimulatedRandomization, an object, representing simulation output.

Result

A vector of variance of imbalance values summarized via simulations.

Expected maximum imbalance over first allocation steps

Incertus.calc_expected_max_abs_imb — Method

Function calculates expected maximum absolute imbalance over first allocations vs. allocation step.

Call

calc_expected_max_abs_imb(sr)

Arguments

sr::SimulatedRandomization: an instance of SimulatedRandomization, an object, representing simulation output.

Result

A vector of expected maximum absolute imbalance over firat allocations values summarized via simulations.

A cumulative average loss at the first allocation steps

Incertus.calc_cummean_loss — Method

Function calculates cumulative average loss vs. allocation step.

Call

calc_expected_abs_imb(sr)

Arguments

sr::SimulatedRandomization: an instance of SimulatedRandomization, an object, representing simulation output.

Result

A vector of cumulative average loss values summarized via simulations.

Cumulative average of expected proportions of correct guesses over first allocation steps under two different guessing strategies

Incertus.calc_cummean_epcg — Method

Function calculates cumulative averages of expected proportions of correct guesses vs. allocation step.

Call

calc_cummean_epcg(sr, gs)

Arguments

sr::SimulatedRandomization: an instance of SimulatedRandomization, an object,

representing simulation output.

gs::String: guessing strategy; accepts two values: "C" (corresponds to the

convergence guessing strategy) of "MP" (corresponds to the maximum probability guessing strategy).

Result

A vector of cumulative averages of the expected proportions of correct guesses values summarized via simulations.

Cumulative average of expected proportions of deterministic assignments over first allocation steps

Incertus.calc_cummean_pda — Method

Function calculates cumulative averages of the proportions of deterministic assignments vs. allocation step.

Call

calc_cummean_pda(sr)

Arguments

sr::SimulatedRandomization: an instance of SimulatedRandomization, an object,

representing simulation output.

Result

A vector of the cumulative averages of the proportions of deterministic assignmnets values summarized via simulations.

Forcing index

Incertus.calc_fi — Method

Function calculates forcing index vs. allocation step.

Call

calc_fi(sr)

Arguments

sr::SimulatedRandomization: an instance of SimulatedRandomization, an object,

representing simulation output.

Result

A vector of the forcing index values summarized via simulations.

Balance-randomness trade-off

Incertus.calc_brt — Method

Function calculates balance-randomness trade-off vs. allocation step.

Call

calc_brt(sr)

Arguments

sr::SimulatedRandomization: an instance of SimulatedRandomization, an object,

representing simulation output.

Result

A vector of balance-randomness trade-off measurement values summarized via simulations.

Visualizing operational characteristics

Incertus.plot — Method

Function used to visualize simulated operational characteristics.

Call

plot(op; kw...)

Arguments

op::DataFrame – a data frame, containing simulated operational characteristics; each column represents simulated output for a particular randomization procedure.

Key words

kw refers to the kew words. The following are currently supported and have to be provided:

xlabel::String – a text label for $X$ axis.

ylabel::String – a text label for $Y$ axis.

colors::Matrix{Symbol} – a matrix (row) of colors (symbolic representation) to distinguish outputs for different randomization procedure.

shapes::Matrix{Symbol} – a matrix (row) of markers (symbolic representation) to distinguish outputs for different randomization procedure.

marker_size::Int64 – marker size.

xticks::Vector{<:Number} – ticks for $X$ axis.

yticks::Vector{<:Number} – ticks for $Y$ axis.

xlims::Tuple{<:Number, <:Number} – a tuple to set limits of $X$ axis; default is set to (0, nrow(op)+1).

ylims::Tuple{<:Number, <:Number} – a tuple to set limits of $Y$ axis; default is set to (0, ceil(maximum([maximum(row) for row in eachrow(op)]))).

size::Tuple{Int64, Int64} – an output figure size; default is set to (800, 600).

dpi::Int64 – dpi value; default is set to 300.

Result

A plot of corresponding operational chacteristics vs. allocation step

Incertus.heatmap — Method

Function used to visualize balance-randomness trade-off as a heatmap plot.

Call

heatmap(op; kw...)

Arguments

brt::DataFrame – a data frame, containing simulated balance-randomness trade-off; each column represents simulated output for a particular randomization procedure.

Key words

kw refers to the kew words. The following are currently supported and have to be provided:

xlabel::String – a text label for $X$ axis; default is set ot "allocation step".

ylabel::String – a text label for $Y$ axis; default is set ot "design".

palette::Vector{Symbol} – a vector of colors (symbolic representation) to make a continuous color gradient; default is set to [:red, :orange, :yellow, :green, :blue, :navy, :purple].

xticks::Vector{<:Number} – ticks for $X$ axis.

size::Tuple{Int64, Int64} – an output figure size; default is set to (800, 600).

dpi::Int64 – dpi value; default is set to 300.

Result

A heatmap plot of the balance-randomness trade-off vs. allocation step

Auxiliary functions

Incertus.set_label — Method

Function sets a label for a randomization procedure.

An example:

julia> w = [1, 2, 3, 4]
4-element Vector{Int64}:
 1
 2
 3
 4

julia> dlr_proc = DLR(w, 2)
DLR(2): restricted randomization procedure, targeting 1:2:3:4 allocation in 4-arm trial.

julia> label = set_label(dlr_proc)
"DLR(2)"