Operational characteristics (two-arm trial with equal 1:1 allocation)

Several measures of imbalance and randomness have been implemented in the package:

Measures of imbalance:

$D(n)$ – final imbalance after all treatment assignments made.

$\mathbf{E}\left[|D(j)|\right]$ – expected absolute imbalance at the $j^\text{th}$ allocation step.

$\mathbf{E}\left[|D(j)|^2\right]=\mathbf{var}\left[D(j)\right]$ – variance of imbalance at the $j^\text{th}$ allocation step.

$\mathbf{E}\left[\max\limits_{1\leq m \leq j}|D(m)|\right]$ – expected maximum imbalance over the first $j$ allocation steps.

$Imb(j) = \frac{1}{j}\sum\limits_{m=1}^j\frac{\mathbf{E}\left[|D(m)|^2\right]}{m}$ – a cumulative average loss at the $j^\text{th}$ allocation step.

Measures of randomness:

$EPCG_{conv}(j) = \frac{1}{j}\sum\limits_{m=1}^j\mathbf{E}\left[G_m = \delta_m\right]$ – cumulative average of expected proportions of correct guesses over first ($j$) allocation steps under the convergence guessing strategy, where $G_m$ is a random variable taking values based on the investigator's guess (if $G_m = \delta_m$, the guess is correct) at the $m^\text{th}$ allocation step, given current imbalance, $D(m-1)$:

\[G_m = \left\{ \begin{array}{rl} 1, & D(m-1) < 0; \\ \sim Bernoulli(0.5), & D(m-1) = 0; \\ 0, & D(m-1) > 0. \end{array} \right. \]

$EPCG_{max}(j) = \frac{1}{j}\sum\limits_{m=1}^j\mathbf{E}\left[\widetilde{G}_m = \delta_m\right]$ – cumulative average of expected proportions of correct guesses over first ($j$) allocation steps under the maximum probability guessing strategy, where $\widetilde{G}_m$ is a random variable taking values based on the investigator's guess (if $G_m = \delta_m$, the guess is correct) at the $m^\text{th}$ allocation step, given allocation probability, $\phi_{m}$:

\[\widetilde{G}_m = \left\{ \begin{array}{rl} 1, & \phi_m > 0.5; \\ \sim Bernoulli(0.5), & \phi_m = 0.5; \\ 0, & \phi_m < 0.5. \end{array} \right. \]

$PD(j) = \frac{1}{j}\sum\limits_{m=1}^j\Pr(\phi_m\in\{0, 1\})$ – cumulative average of expected proportions of deterministic assignments over first ($j$) allocation steps.

$FI(j) = \frac{4}{j}\sum\limits_{m=1}^j\mathbf{E}\left[|\phi_m-0.5|\right]$ – forcing index, which takes values on a scale 0–1. $FI(j) = 0, \forall j$ for CRD and $FI(j) = 1$ for PBD with a block size $bs=2$, assuming $j$ is even (most balanced design).

Balance-randomness trade-off:

$G(j) = \sqrt{\left\{Imb(j)\right\}^2 + \left\{FI(j)\right\}^2}$ represents a balance-randomness trade-off at the $j^\text{th}$ allocation step. Lower values of $G(j)$ indicate better balance–randomness trade-off.

Below, there are functions available for calculating operational characteristics.

Final imbalance

Incertus.calc_final_imb — Method

Function calculates final imbalance

Call

calc_final_imb(sr)

Arguments

sr::SimulatedRandomization: an instance of SimulatedRandomization, an object, representing simulation output.

Result

A Vector of final imbalance values obtained via simulations.

Incertus.calc_final_imb — Method

Function calculates final imbalance

Call

calc_final_imb(sr)

Arguments

sr::Vector{SimulatedRandomization}: a vector of instances of SimulatedRandomization, representing simulation output.

Result

A DataFrame of simulated final imbalances' values obtained via simulations.

Expected absolute imbalance

Incertus.calc_expected_abs_imb — Method

Function calculates expected absolute imbalance vs. allocation step.

Call

calc_expected_abs_imb(sr)

Arguments

sr::SimulatedRandomization: an instance of SimulatedRandomization, an object, representing simulation output.

Result

A Vector of expected absolute imbalance values summarized via simulations.

Incertus.calc_expected_abs_imb — Method

Function calculates expected absolute imbalance vs. allocation step.

Call

calc_expected_abs_imb(sr)

Arguments

sr::Vector{SimulatedRandomization}: a vector of instances of SimulatedRandomization, representing simulation output.

Result

A DataFrame of expected absolute imbalances' values summarized via simulations.

Variance of imbalance

Incertus.calc_variance_of_imb — Method

Function calculates variance of imbalance vs. allocation step.

Call

calc_variance_of_imb(sr)

Arguments

sr::SimulatedRandomization: an instance of SimulatedRandomization, an object, representing simulation output.

Result

A Vector of variance of imbalance values summarized via simulations.

Incertus.calc_variance_of_imb — Method

Function calculates variance of imbalance vs. allocation step.

Call

calc_variance_of_imb(sr)

Arguments

sr::Vector{SimulatedRandomization}: a vector of instances of SimulatedRandomization, representing simulation output.

Result

A DataFrame of variances' of imbalance values summarized via simulations.

Expected maximum imbalance over first allocation steps

Incertus.calc_expected_max_abs_imb — Method

Function calculates expected maximum absolute imbalance over first allocations vs. allocation step.

Call

calc_expected_max_abs_imb(sr)

Arguments

sr::SimulatedRandomization: an instance of SimulatedRandomization, an object, representing simulation output.

Result

A Vector of expected maximum absolute imbalance over firat allocations values summarized via simulations.

Incertus.calc_expected_max_abs_imb — Method

Function calculates expected maximum absolute imbalance over first allocations vs. allocation step.

Call

calc_expected_max_abs_imb(sr)

Arguments

sr::Vector{SimulatedRandomization}: a vector of instances of SimulatedRandomization, representing simulation output.

Result

A DataFrame of expected maximum absolute imbalances' values (over first allocations) summarized via simulations.

A cumulative average loss at the first allocation steps

Incertus.calc_cummean_loss — Method

Function calculates cumulative average loss vs. allocation step.

Call

calc_cummean_loss(sr)

Arguments

sr::SimulatedRandomization: an instance of SimulatedRandomization, an object, representing simulation output.

Result

A Vector of cumulative average loss' values summarized via simulations.

Incertus.calc_cummean_loss — Method

Function calculates cumulative average loss vs. allocation step.

Call

calc_cummean_loss(sr)

Arguments

sr::Vector{SimulatedRandomization}: a vector of instances of SimulatedRandomization, representing simulation output.

Result

A DataFrame of cumulative average losses' values summarized via simulations.

Cumulative average of expected proportions of correct guesses over first allocation steps under two different guessing strategies

Incertus.calc_cummean_epcg — Method

Function calculates cumulative averages of expected proportions of correct guesses vs. allocation step.

Call

calc_cummean_epcg(sr, gs)

Arguments

sr::SimulatedRandomization: an instance of SimulatedRandomization, an object, representing simulation output.

gs::String: guessing strategy; accepts two values: "C" (corresponds to the convergence guessing strategy) of "MP" (corresponds to the maximum probability guessing strategy).

Result

A Vector of cumulative averages of the expected proportions of correct guesses summarized via simulations.

Incertus.calc_cummean_epcg — Method

Function calculates cumulative averages of expected proportions of correct guesses vs. allocation step.

Call

calc_cummean_epcg(sr, gs)

Arguments

sr::Vector{SimulatedRandomization}: a vector of instances of SimulatedRandomization, representing simulation output.

gs::String: guessing strategy; accepts two values: "C" (corresponds to the convergence guessing strategy) of "MP" (corresponds to the maximum probability guessing strategy).

Result

A vector of cumulative averages of the expected proportions of correct guesses summarized via simulations.

Cumulative average of expected proportions of deterministic assignments over first allocation steps

Incertus.calc_cummean_pda — Method

Function calculates cumulative averages of the proportions of deterministic assignments vs. allocation step.

Call

calc_cummean_pda(sr)

Arguments

sr::SimulatedRandomization: an instance of SimulatedRandomization, an object, representing simulation output.

Result

A vector of the cumulative averages of the proportions of deterministic assignmnets values summarized via simulations.

Incertus.calc_cummean_pda — Method

Function calculates cumulative averages of the proportions of deterministic assignments vs. allocation step.

Call

calc_cummean_pda(sr)

Arguments

sr::Vector{SimulatedRandomization}: a vector of instances of SimulatedRandomization, representing simulation output.

Result

A DataFrame of the cumulative averages of the proportions of deterministic assignmnets summarized via simulations.

Forcing index

Incertus.calc_fi — Method

Function calculates forcing index vs. allocation step.

Call

calc_fi(sr)

Arguments

sr::SimulatedRandomization: an instance of SimulatedRandomization, an object, representing simulation output.

Result

A Vector of the forcing index values summarized via simulations.

Incertus.calc_fi — Method

Function calculates forcing index vs. allocation step.

Call

calc_fi(sr)

Arguments

sr::Vector{SimulatedRandomization}: a vector of instances of SimulatedRandomization, representing simulation output.

Result

A DataFrame of the forcing index values summarized via simulations.

Balance-randomness trade-off

Incertus.calc_brt — Method

Function calculates balance-randomness trade-off vs. allocation step.

Call

calc_brt(sr)

Arguments

sr::SimulatedRandomization: an instance of SimulatedRandomization, an object,

representing simulation output.

Result

A Vector of balance-randomness trade-off measurement values summarized via simulations.

Incertus.calc_brt — Method

Function calculates balance-randomness trade-off vs. allocation step.

Call

calc_brt(sr)

Arguments

sr::Vector{SimulatedRandomization}: a vector of instances of SimulatedRandomization, representing simulation output.

Result

A DataFrame of balance-randomness trade-off measurements' values summarized via simulations.