Visualizing operational characteristics

Incertus.plot — Method

Function used to visualize simulated operational characteristics.

Call

plot(op; kwargs...)

Arguments

op::DataFrame – a data frame, containing simulated operational characteristics; each column represents simulated output for a particular randomization procedure.

kwargs refers to the kew words. Here, it is possible to pass key-value pairs supported by a StatsPlots.heatmap function.

Result

A plot of corresponding operational chacteristics vs. allocation step

Incertus.heatmap — Method

Function used to visualize balance-randomness trade-off as a heatmap plot.

Call

heatmap(brt; kwargs...)

Arguments

brt::DataFrame – a data frame, containing simulated balance-randomness trade-off; each column represents simulated output for a particular randomization procedure.

kwargs refers to the kew words. Here, it is possible to pass key-value pairs supported by a StatsPlots.heatmap function.

Result

A heatmap plot of the balance-randomness trade-off vs. allocation step

Incertus.violin — Method

Function used to visualize a distribution of the final imbalance as a violin plot.

Call

violin(final_imb; kwargs...)

Arguments

final_imb::DataFrame – a data frame, containing simulated final imbalances; each column represents simulated output for a particular randomization procedure.

kwargs refers to the kew words. Here, it is possible to pass key-value pairs supported by a StatsPlots.violin function.

Result

A violin plot of the final imbalances.